backend/requirements.txt
fastapi
uvicorn
google-generativeai
replicate
requests
python-multipart
python-dotenv

